+++ main-site/components/DiscourseTopicFact.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable @next/next/no-img-element */
 import React from 'react';
 import { formatFactDate } from '~/lib/date';
-import { Fact } from '~/lib/groq/fact.partial.groq';
-import { Picture } from '~/lib/groq/picture.partial.groq';
+import { Fact } from '~/lib/groq/fact.projection';
+import { Picture } from '~/lib/groq/picture.projection';
 import { sanityImage } from '~/shared/lib/sanityio';
 
 export function DiscourseTopicFact(
