+++ main-site/components/Fact.tsx
@@ -3,8 +3,8 @@ import React, { PropsWithChildren } from 'react';
 import { BiHash, BiLink, BiMessage } from 'react-icons/bi';
 import { Tag } from '~/components/Tag';
 import { getSourceHost } from '~/lib/getSourceHost';
-import { Celeb } from '~/lib/groq/celeb.partial.groq';
-import { Fact as TFact } from '~/lib/groq/fact.partial.groq';
+import { Celeb } from '~/lib/groq/celeb.projection';
+import { Fact as TFact } from '~/lib/groq/fact.projection';
 import { Link } from '~/lib/Link';
 import Image from 'next/image';
 import { FaQuoteLeft } from 'react-icons/fa';
