+++ main-site/lib/groq/fact.projection.ts
