+++ main-site/lib/groq/picture.projection.ts
