+++ main-site/pages/index.tsx
@@ -12,8 +12,8 @@ import { Page } from '~/components/Page';
 import { Spinner } from '~/components/Spinner';
 import { formatFactDate } from '~/lib/date';
 import { getTrendingCelebs, TrendingCelebs } from '~/lib/getTrendingCelebs';
-import { Fact as TFact, factPartialGroq } from '~/lib/groq/fact.partial.groq';
-import { Picture } from '~/lib/groq/picture.partial.groq';
+import { Fact as TFact, factPartialGroq } from '~/lib/groq/fact.projection';
+import { Picture } from '~/lib/groq/picture.projection';
 import { Link } from '~/lib/Link';
 import { log } from '~/shared/lib/log';
 import { sanityClient } from '~/shared/lib/sanityio';
